Rename 'trait' package to 'traits' to satisfy Idea
Idea doesn't like package named 'trait'.
